api/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pgvector
alembic
typer[all]
bcrypt
pytest
httpx